Refactored touch listener interfaces.

+++ src/edu/berkeley/cellscope/cscore/cameraui/ZoomablePreview.java
@@ -2,6 +2,6 @@ package edu.berkeley.cellscope.cscore.cameraui;
 
 public interface ZoomablePreview extends TouchControllable {
 
-	public double getMaxZoom();
+	public int getMaxZoom();
 	public void zoom(int amount);
 }
